requirements.txt
pillow >= 3.0.0
numpy >= 1.9.2
scipy >= 0.16.1
